refactor: input to io handler
